Modificada las actividades para recibir valores con coma

+++ CarbonCalcutaorBackEnd/src/main/java/net/atos/zerokhoi/entity/Actividad.java
@@ -44,7 +44,7 @@ public class Actividad implements Serializable{
 	
 	@Column(nullable= false)
 	@NotEmpty
-	private Long consumo;
+	private Double consumo;
 	
 	
 	@ManyToOne()
